requirements.txt
tensorflow==2.15.0
streamlit==1.36.0
pillow==10.3.0
numpy==1.26.4
matplotlib==3.9.2
opencv-python==4.8.1.78
scikit-learn==1.3.0
pandas==2.1.2
tqdm==4.66.1
